### 2020-05-21 #2200 @kthyng
Hi @oliviaguest! You were the suggested handling editor for this submission. I'm hoping to get some extra eyes on this submission being in scope given that the number of lines of code shown above is a bit low for a typical submission (being below 1000 for the number of lines of R code).

Given all this, are you able to edit this submission? Thanks!
