+++ man/geom_flat_violin.Rd
@@ -4,9 +4,17 @@
 \alias{geom_flat_violin}
 \title{ggplot Flat Violin}
 \usage{
-geom_flat_violin(mapping = NULL, data = NULL, stat = "ydensity",
-  position = "dodge", trim = TRUE, scale = "area",
-  show.legend = NA, inherit.aes = TRUE, ...)
+geom_flat_violin(
+  mapping = NULL,
+  data = NULL,
+  stat = "ydensity",
+  position = "dodge",
+  trim = TRUE,
+  scale = "area",
+  show.legend = NA,
+  inherit.aes = TRUE,
+  ...
+)
 }
 \arguments{
 \item{mapping}{A value}
