+++ DESCRIPTION
@@ -16,7 +16,6 @@ Depends: dplyr,
          rlang
 Imports: data.table,
          fda,
-         fractal,
          lazyeval,
          mgcv,
          signal,
@@ -24,4 +23,4 @@ Imports: data.table,
          tidyr,
          utils,
          zoo
-RoxygenNote: 6.1.1
+RoxygenNote: 7.1.0
