### 2020-05-26 68bf490
Update plot_pupil_data.R
replace fun.y with fun in ggplot2::stat_summary
