### 2020-06-01 #2285 @whedon
[REVIEW]: PupillometryR: An R package for preparing and analysing pupillometry data
**Submitting author:** @samhforbes (<a href="http://orcid.org/0000-0003-1022-4676">Samuel Forbes</a>)
**Repository:** <a href="https://github.com/samhforbes/PupillometryR" target ="_blank">https://github.com/samhforbes/PupillometryR</a>
**Version:** v0.0.3
