### 2020-06-01 #2285 @oliviaguest
👋  @paulinepalma, @Athanasiamo: Any code-related questions feel free to open issues on the package's repo itself and then link to whatever issue you open from here. Everything important about the PDF or the package, like high-level questions to me or @samhforbes, discussions and feedback, just leave it here as a comment directly. ✨ 🌷
