### 2020-06-02 samhforbes/PupillometryR#6 @Athanasiamo
Move update information to a NEWS file
Small descriptions of version updates should ideally be ni a `NEWS` file, to not start cluttering the README as development progresses.
