### 2020-06-02 samhforbes/PupillometryR#7 @Athanasiamo
insufficient licence file
The licence file contains no information about the type of licence or description of use, derivative work distribution etc.

Please apply an [OSI approved licence](https://opensource.org/licenses/alphabetical). this can for instance easily be done though `usethis::use_xx_licence()`
