### 2020-06-02 samhforbes/PupillometryR#8 @Athanasiamo
CRAN installation instructions
Its great your package is on CRAN! But the readme does not say so, and I had to read the paper to know that it was. Please add information on how to install from CRAN on the main README page
