### 2020-06-02 samhforbes/PupillometryR#9 @Athanasiamo
Move most readme content to a vignette
This is an opinion, so it can absolutely be discussed.

in my opinion, the readme as a whole should be moved into a `vignette` folder and be compiled as such. This will also then run the vignette code during build checks, making sure that the vignette is actually executable  (easiest way to do so it to use `usethis::use_vignette()`  ).

The readme should be made concise and rather highlight what the package intention and core features are, while a full user tutorial is best placed in a vignette.
