### 2020-06-02 samhforbes/PupillometryR#10 @Athanasiamo
Sylvains tutorial
There is a mention of a tutorial by Sylvain in the README, this should be linked to for easy reference.
