### 2020-06-02 samhforbes/PupillometryR#11 @Athanasiamo
Add community guidelines
There is currently no information on:
1. how to contribute
2. how to make a bug report
3. how to seek support

Please add some simple instructions to the README or package DESCRIPTION this.

The DESCRIPTION for instance allows for a `BugReports:` field
