### 2020-06-02 #2285 @Athanasiamo
# Review checklist for @Athanasiamo
## Conflict of interest
✅    I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

## Code of Conduct
✅  I confirm that I read and will adhere to the JOSS code of conduct.

## General checks
 ✅  Repository: Is the source code for this software available at the repository url?
 ❌  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
    - LICENCE file does not contain any information on the type of licence to apply to the package.
    - recommendation: utilise `usethis::use_[xx]_licence` to create a valid licence file.
    - Noted in [this isse](https://github.com/samhforbes/PupillometryR/issues/7)
✅  Contribution and authorship: Has the submitting author (@samhforbes) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
    - One other contributor in a single minor commit. This looks ok to me.
    - Another package author is listed without any commit history. This person is mentioned in the paper acknowledgement.

## Functionality
✅  Installation: Does installation proceed as outlined in the documentation?
✅  Functionality: Have the functional claims of the software been confirmed?
✅  Performance: If there are any performance claims of the software, have they been confirmed?

## Documentation
 ✅   A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
✅  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
    * since the package is also on CRAN, installation instructions should clearly state that, and only point to github installation for development versions.  I had to read the paper to know it was on CRAN. [noted in this issue](https://github.com/samhforbes/PupillometryR/issues/8)
➖  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
    * yes but not in an ideal way. in my opinion.
    * Great with data shipped with the package so one can give the functions a go.
    *  in my opinion, the readme as a whole should be moved into a `vignette` folder and be compiled as such. This will also then run the vignette code during build checks, making sure that the vignette is actually executable  (easiest way to do so it to use `usethis::use_vignette()`  ).  [Noted in this issue](https://github.com/samhforbes/PupillometryR/issues/9)
    * please add link to Sylvains tutorial in the README [Noted in this issue](https://github.com/samhforbes/PupillometryR/issues/10)
    * A new readme just highlighting some core functionality of the package can be used in stead.
    * Instead of informing about patch updates ( like the top of the current readme), use a `NEWS` file instead for this sort of information.  [noted in this issue](https://github.com/samhforbes/PupillometryR/issues/6)
✅  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?  **Good**
✅   Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
    - function examples are run on build checks, and everything checks out fine.
    - But there are no CI checks.
❌  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
    - There is no such information, [as noted in this issue](https://github.com/samhforbes/PupillometryR/issues/11)

## Software paper
✅  Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
✅  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
✅  State of the field: Do the authors describe how this software compares to other commonly-used packages?
✅  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
✅  References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
