### 2020-06-02 samhforbes/PupillometryR#10 @samhforbes
Hi @Athanasiamo thanks for flagging this. As things currently stand it is linked (see lines 52 and 62 of README.md), but with the changes to the README you have suggested that location may yet move.
