### 2020-06-02 1ff5017
Add news.md
Fixing layout with respect to #6
