+++ NEWS.md
@@ -1,0 +1,4 @@
+### Update May 2020 PupillometryR 0.0.2
+
+-   Minor patch to avoid breaks with new version of dplyr
+-   Warning errors removed from ggplot2 when plotting
