+++ README.md
@@ -1,6 +1,3 @@
-### Update May 2020 PupillometryR 0.0.2
-
--   Minor patch to avoid breaks with new version of dplyr
 
 Introduction
 ------------
@@ -25,15 +22,15 @@ library(PupillometryR)
 
     ## Loading required package: dplyr
 
-    ## 
+    ##
     ## Attaching package: 'dplyr'
 
     ## The following objects are masked from 'package:stats':
-    ## 
+    ##
     ##     filter, lag
 
     ## The following objects are masked from 'package:base':
-    ## 
+    ##
     ##     intersect, setdiff, setequal, union
 
     ## Loading required package: ggplot2
@@ -144,7 +141,7 @@ plot(new_data, pupil = LPupil, group = 'condition')
 ![](README_files/figure-markdown_github/unnamed-chunk-7-1.svg)
 
 ``` r
-plot(new_data, pupil = LPupil, group = 'subject') 
+plot(new_data, pupil = LPupil, group = 'subject')
 ```
 
     ## Warning: Removed 3639 rows containing non-finite values (stat_summary).
@@ -170,8 +167,8 @@ Now that we have done that, we want the mean of the two pupil sizes, so
 let’s see how that looks:
 
 ``` r
-mean_data <- calculate_mean_pupil_size(data = regressed_data, 
-                                       pupil1 = RPupil, 
+mean_data <- calculate_mean_pupil_size(data = regressed_data,
+                                       pupil1 = RPupil,
                                        pupil2 = LPupil)
 
 plot(mean_data, pupil = mean_pupil, group = 'subject')
@@ -201,7 +198,7 @@ Now we need to clean up our data - let’s first assess how much missing
 data there is:
 
 ``` r
-missing <- calculate_missing_data(mean_data, 
+missing <- calculate_missing_data(mean_data,
                                   mean_pupil)
 head(missing)
 ```
@@ -209,11 +206,11 @@ head(missing)
     ## # A tibble: 6 x 3
     ##   ID    Trial Missing
     ##   <chr> <fct>   <dbl>
-    ## 1 1     Easy1  0.104 
-    ## 2 1     Hard1  0.144 
+    ## 1 1     Easy1  0.104
+    ## 2 1     Hard1  0.144
     ## 3 1     Easy2  0.0846
-    ## 4 1     Hard2  0.104 
-    ## 5 1     Easy3  0.149 
+    ## 4 1     Hard2  0.104
+    ## 5 1     Easy3  0.149
     ## 6 1     Hard3  0.139
 
 We can see if we view the whole file that participant 6 has a fair
@@ -233,10 +230,10 @@ mean_data2 <- clean_missing_data(mean_data,
                                  subject_trial_threshold = .75)
 ```
 
-    ## Removing trials with a proportion missing > 0.75 
+    ## Removing trials with a proportion missing > 0.75
     ##  ...removed 0 trials
 
-    ## Removing subjects with a proportion of missing trials > 0.75 
+    ## Removing subjects with a proportion of missing trials > 0.75
     ##  ...removed 0 subjects
 
 Now we come to filtering the data. PupillometryR offers 3 filter types:
@@ -342,16 +339,16 @@ We could then simply analyse this with a t-test if we wished.
 t.test(mean_pupil ~ Type, paired = T, data = window)
 ```
 
-    ## 
+    ##
     ##  Paired t-test
-    ## 
+    ##
     ## data:  mean_pupil by Type
[…]
