### 2020-06-02 ecea29c
MIT license
Provided in full here - see #7
