### 2020-06-02 ddb8567
Moce main readme info to vignette
