+++ vignettes/PupillometryR.Rmd
@@ -1,0 +1,343 @@
+---
+title: "PupillometryR"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{PupillometryR}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(PupillometryR)
+```
+
+This package is designed to make dealing with pupil data (perhaps more traditionally done in MATLAB) easier to wrangle in R. It makes heavy use of a few packages which should be acknowledged here, especially the (excellent) packages fda and signal.
+
+As well as the above packages, it is very important to note that the type of analysis shown here has been available in MATLAB for a while, and there is an excellent tutorial on it, which I thoroughly recommend reading first, written by Sylvain Sirois, [here:](https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=).
+
+It's worth making sure that your setup and experiment do facilitate the use of pupillometry - it may not be suited for all kinds of experiment.
+
+## Getting started
+
+We will first run through an example analysis with the data provided in the package, which, again, comes from Sylvain Sirois' tutorial on [his webpage](https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=).
+
+The first thing I would recommend doing is having a close look at the pupil data. Eyetrackers have a couple of different ways of dealing with this, so it's important to know a few things:
+
+1. What unit of measurement is being used
+2. What value is given to missing data or blinks (typically ., -1, or *NA*)
+3. What framerate were you recording at, and is this consistent.
+
+The data here is an eyetracking experiment with hard and easy trials, performed on 9 participants. Participant 8 needs to be removed (more details on Sylvain's tutorial).
+
+```{r}
+data("pupil_data")
+
+#Check that IDs are not numeric
+pupil_data$ID <- as.character(pupil_data$ID)
+#remove participant number 8, who had problematic data
+pupil_data <- subset(pupil_data, ID != 8)
+#blinks were registered as -1, so replace with NAs
+pupil_data$LPupil[pupil_data$LPupil == -1] <- NA
+pupil_data$RPupil[pupil_data$RPupil == -1] <- NA
+
+```
+
+The plotting is also a theme of this tutorial, so I will set a nice theme that makes the plots look pretty:
+
+```{r}
+library(ggplot2)
+theme_set(theme_classic(base_size = 12))
+```
+
+First up, we need to put the data into pupillometryR format for further analysis
+
+```{r}
+Sdata <- make_pupillometryr_data(data = pupil_data,
+                                 subject = ID,
+                                 trial = Trial,
+                                 time = Time,
+                                 condition = Type)
+```
+
+In the current data, it is not a concern, but there may be a situation where certain timebins are missing from your data. This can be fixed here, and we will look at the raw data:
+
+```{r}
+new_data <- replace_missing_data(data = Sdata)
+
+head(new_data)
+```
+
[…]
