+++ DESCRIPTION
@@ -24,3 +24,7 @@ Imports: data.table,
          utils,
          zoo
 RoxygenNote: 7.1.0
+Suggests: 
+    knitr,
+    rmarkdown
+VignetteBuilder: knitr
