### 2020-06-02 samhforbes/PupillometryR#11 @samhforbes
Thanks for flagging this! The description currently has a BugReports field, but this will be made clearer in the new README file as well as the other two being made explicit.
