### 2020-06-02 4fc1d15
Update PupillometryR.Rmd
avoid referring to unneeded packages  in the vignette
