### 2020-06-02 8fba8be
add itsadug to vignette
