### 2020-06-02 #2285 @oliviaguest
@openjournals/joss-editors @openjournals/joss-eics has anybody seen a case where the reviewer cannot tick/check the boxes above? @Athanasiamo is not able to interact with the OP, is there a simple solution we are not aware of? She is logged in and I have sent her a video of what it looks like from my end. I have no idea what is wrong — thanks.
