### 2020-06-02 #2285 @paulinepalma
Hi Olivia,

I think I have the same problem as Athanasia. Could you please send me the
invite again? Sorry about that!

Pauline

Le mar. 2 juin 2020, à 09 h 53, whedon <notifications@github.com> a écrit :

> The reviewer already has a pending invite.
>
> @Athanasiamo <https://github.com/Athanasiamo> please accept the invite by
> clicking this link:
> https://github.com/openjournals/joss-reviews/invitations
>
