### 2020-06-02 #2285 @danielskatz
@oliviaguest - if the invite has expired, the whedon re-invite command can also be used, but in this case, it seems unlike that the invite would have expired :)
