### 2020-06-02 #2285 @paulinepalma
Hi Daniel,

That is what I did initially, but here is the message I got: "Sorry, we
couldn't find that repository invitation. It is possible that the
invitation was revoked or that you are not logged into the invited
account." I am logged into my github account.

Pauline

Le mar. 2 juin 2020, à 10 h 27, Daniel S. Katz <notifications@github.com> a
écrit :

> @oliviaguest <https://github.com/oliviaguest> - if the invite has
> expired, the whedon re-invite command can also be used, but in this case,
> it seems unlike that the invite would have expired :)
>
