### 2020-06-02 99f6bf8
update README
move details to Vignette. Should address #8 #9 #10 #11
