+++ README.Rmd
@@ -1,0 +1,98 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# PupillometryR
+
+<!-- badges: start -->
+[![CRAN status](https://www.r-pkg.org/badges/version/PupillometryR)](https://CRAN.R-project.org/package=PupillometryR)
+[![Travis build status](https://travis-ci.org/samhforbes/PupillometryR.svg?branch=master)](https://travis-ci.org/samhforbes/PupillometryR)
+<!-- badges: end -->
+
+The goal of PupillometryR is to to pre-process and then analyze simple pupil
+experiments in R.
+
+## Installation
+
+You can install the released version of PupillometryR from [CRAN](https://CRAN.R-project.org) with:
+
+``` r
+install.packages("PupillometryR")
+```
+
+And the development version from [GitHub](https://github.com/) with:
+
+``` r
+# install.packages("devtools")
+devtools::install_github("samhforbes/PupillometryR")
+```
+## Setup
+
+This package (and the example dataset) was designed in part, based on Sylvain Sirois' MATLAB tutorial, [which can be found here](https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=).
+
+To use all the functionality and plots that follow from the PupillometryR pipeline, please start with *make_pupillometryr_data*, e.g.:
+
+```{r example}
+library(PupillometryR)
+
+data("pupil_data")
+
+#Check that IDs are not numeric
+pupil_data$ID <- as.character(pupil_data$ID)
+#remove participant number 8, who had problematic data
+pupil_data <- subset(pupil_data, ID != 8)
+#blinks were registered as -1, so replace with NAs
+pupil_data$LPupil[pupil_data$LPupil == -1] <- NA
+pupil_data$RPupil[pupil_data$RPupil == -1] <- NA
+
+Sdata <- make_pupillometryr_data(data = pupil_data,
+                                 subject = ID,
+                                 trial = Trial,
+                                 time = Time,
+                                 condition = Type)
+```
+
+All further functions associated with the package follow from there. 
+For example:
+
+```{r}
+plot(Sdata, pupil = LPupil)
+
+```
+
+To follow a detailed walkthrough, run:
+
+```{r eval = F}
+vignette('PupillometryR')
+```
+
+## Getting help
+
+Please use the issues tab (https://github.com/samhforbes/PupillometryR/issues) to file any bugs or suggestions.
[…]
