+++ README.md
@@ -1,77 +1,59 @@
 
-Introduction
-------------
+<!-- README.md is generated from README.Rmd. Please edit that file -->
 
-PupillometryR is a package to pre-process and then analyze simple pupil
-experiments in R. The best way to download it is to open up R in
-RStudio, and download it after [downloading
-devtools](https://www.r-project.org/nosvn/pandoc/devtools.html). It
-first needs to be downloaded before running, if you have devtools, it
-can be done directly from the console in R:
+# PupillometryR
+
+<!-- badges: start -->
+
+[![CRAN
+status](https://www.r-pkg.org/badges/version/PupillometryR)](https://CRAN.R-project.org/package=PupillometryR)
+[![Travis build
+status](https://travis-ci.org/samhforbes/PupillometryR.svg?branch=master)](https://travis-ci.org/samhforbes/PupillometryR)
+<!-- badges: end -->
+
+The goal of PupillometryR is to to pre-process and then analyze simple
+pupil experiments in R.
+
+## Installation
+
+You can install the released version of PupillometryR from
+[CRAN](https://CRAN.R-project.org) with:
 
 ``` r
-devtools::install_github('samhforbes/PupillometryR')
+install.packages("PupillometryR")
 ```
 
-Once we have the package downloaded, we need to load in the package to
-get started:
+And the development version from [GitHub](https://github.com/) with:
+
+``` r
+# install.packages("devtools")
+devtools::install_github("samhforbes/PupillometryR")
+```
+
+## Setup
+
+This package (and the example dataset) was designed in part, based on
+Sylvain Sirois’ MATLAB tutorial, [which can be found
+here](https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=).
+
+To use all the functionality and plots that follow from the
+PupillometryR pipeline, please start with *make\_pupillometryr\_data*,
+e.g.:
 
 ``` r
 library(PupillometryR)
-```
+#> Loading required package: dplyr
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
+#> Loading required package: ggplot2
+#> Loading required package: rlang
 
-    ## Loading required package: dplyr
-
-    ##
-    ## Attaching package: 'dplyr'
-
-    ## The following objects are masked from 'package:stats':
-    ##
-    ##     filter, lag
-
-    ## The following objects are masked from 'package:base':
-    ##
-    ##     intersect, setdiff, setequal, union
-
-    ## Loading required package: ggplot2
-
-    ## Loading required package: rlang
-
-This package is designed to make dealing with pupil data (perhaps more
-traditionally done in MATLAB) easier to wrangle in R. It makes heavy use
-of a few packages which should be acknowledged here, especially the
-(excellent) packages fda and signal.
-
-As well as the above packages, it is very important to note that the
-type of analysis shown here has been available in MATLAB for a while,
-and there is an excellent tutorial on it, which I thoroughly recommend
-reading first, written by Sylvain Sirois,
-[here:](https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=).
-
-It’s worth making sure that your setup and experiment do facilitate the
[…]
