### 2020-06-02 samhforbes/PupillometryR#9 @samhforbes
I really appreciate the benefit of your experience here @Athanasiamo. As you have suggested, the README has been simplified, with the expanded walkthrough moved to vignettes. In line with the other issues you've raised, the other tutorials are linked, community guidelines are flagged, and installation instructions are (hopefully) clear - more in line with the usethis templates. I have included badges to note the travis-CI and CRAN status as well.
