### 2020-06-02 #2285 @paulinepalma
@oliviaguest I started going through the repository and I am having difficulties installing the developper version of the package (under R version 4.0.0). I ended up installing the CRAN version, which works. It could be an issue with my own computer, so I am unsure whether I should mention this in the issues?
