### 2020-06-02 samhforbes/PupillometryR#12 @paulinepalma
Developer version install
I am having difficulties with the developer version of the package (running on R 4.0.0 on mac os 10.12.6).
Basically, I ran the following line:
devtools::install_github("samhforbes/PupillometryR")

And I ran into the following error:

Downloading GitHub repo samhforbes/PupillometryR@master
✓  checking for file ‘/private/var/folders/3q/4p41_jds5gdd_g0kjjtgvsnm0000gp/T/Rtmp8JVrje/remotes342215f87e9/samhforbes-PupillometryR-67c292b/DESCRIPTION’ (1.2s)
─  preparing ‘PupillometryR’:
✓  checking DESCRIPTION meta-information ...
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in  serialize/load version 3 cannot be read in older versions of R.  File(s) containing such objects:  ‘PupillometryR/data/pupil_data.RData’
─  building ‘PupillometryR_0.0.2.9000.tar.gz’

dyld: lazy symbol binding failed: Symbol not found: _utimensat
  Referenced from: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libR.dylib (which was built for Mac OS X 10.13)
  Expected in: /usr/lib/libSystem.B.dylib

dyld: Symbol not found: _utimensat
  Referenced from: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libR.dylib (which was built for Mac OS X 10.13)
  Expected in: /usr/lib/libSystem.B.dylib

/Library/Frameworks/R.framework/Resources/bin/INSTALL: line 34:  1251 Done                    echo 'tools:::.install_packages()'
      1252 Abort trap: 6           | R_DEFAULT_PACKAGES= LC_COLLATE=C "${R_HOME}/bin/R" $myArgs --no-echo --args ${args}
Error: Failed to install 'PupillometryR' from GitHub:
  (converted from warning) installation of package ‘/var/folders/3q/4p41_jds5gdd_g0kjjtgvsnm0000gp/T//Rtmp8JVrje/file3426d3e31d/PupillometryR_0.0.2.9000.tar.gz’ had non-zero exit status

My OS version is pretty old, so I think the problem may be on my end. I had no issue with the released version. Although the line "vignette('PupillometryR')" from the Readme doc does not work (not sure if the vignette is included in the released version). I had to copy/paste the code from the vignette into R.
