### 2020-06-02 samhforbes/PupillometryR#13 @paulinepalma
Paper typos
There are a few typos in the paper itself:

Paragraph 2: “The first, is that…”
Remove comma (subject never separated from verb)

Paragraph 2: “an available for usage with most common brands of eye-tracker”
an**d** available for usage
