### 2020-06-02 samhforbes/PupillometryR#14 @paulinepalma
References/citations
**Paper**
The citation for the fda package is Ramsay, Hooker and Grave (2020), not Ramsay & Silverman (1997).

**Readme file**
In the "Getting help" section, the FDA website triggers a security warning. I recommend indicating the publication date of Ramsay and Silverman book. I believe the first author's name is spelled "Ramsay", not "Ramsey". The link pointing to Allen et al. is incorrect, it links to Eadie & Kapsner-Smith (2011). The acronyms GAMMs and GAMs are used interchangeably, I recommend sticking to one term.
In the References section, Jackson & Sirois (2009) is not mentioned in the text but it is listed here. It is actually in the vignette, so maybe it should be kept here.
Please include a link to the Allen et al. paper.
