### 2020-06-02 #2285 @oliviaguest
@paulinepalma @Athanasiamo thanks for both your reviews so far! Super helpful and I really appreciate the work you are putting in. 👏

I edited your comments. I hope/think I left only the ones with relevant info over and above the tick (if I removed anything vital, please feel free to put it back — and sorry if so!). TBC it's to help @samhforbes focus on what he needs to address. The stuff that just has a ✅ is already in the OP. I hope this makes sense!
