### 2020-06-03 samhforbes/PupillometryR#12 @samhforbes
Hi @paulinepalma, this is a tricky one. It looks like it might be an issue further up the food chain. I have a few thoughts that might be a help:
1. does devtools::install_github() work with any other repository? You can check if it works with devtools::has_devel() - possibly related to issue 3 below.
2. R 4.0.0 looks like it only supports macOS 10.13 upwards (I don't know if you've encountered any further issues with this?
3. (I think this might be most likely) it could be an issue with XCode, which I think didn't ship with OX 10.12.

Please let me know if any of these help.
