### 2020-06-03 76d61f6
Update paper.md
Fix to #13
