### 2020-06-03 7c6059f
Update references
Fix for #14 .
