+++ paper/paper.bib
@@ -37,6 +37,14 @@
     keywords = {Baseline correction, Pupil size, Pupillometry, Research methods}
 }
 
+@misc{Ramsay2020,
+author = {Ramsay, J. O. and {Graves S.} and Hooker, G.},
+title = {{fda: Functional Data Analysis}},
+year = {2020}
+url = {https://CRAN.R-project.org/package=fda }
+}
+
+
 @article{Hershman2019,
     title = {{CHAP: Open-source software for processing and analyzing pupillometry data}},
     year = {2019},
