### 2020-06-03 samhforbes/PupillometryR#14 @samhforbes
Hi @paulinepalma this should be fixed now - I clearly redid the README in too much haste! Thanks for noticing.
