### 2020-06-03 samhforbes/PupillometryR#12 @paulinepalma
Hi @samhforbes, thank you for the suggestions! I tried looking more into this:

1. The problem does not look like it comes from devtools, as devtools::has_devel() returns the message "Your system is ready to build packages!".
2. I updated to R 4.0.0 recently and did not run into any issue so far.
3. I don't think I have XCode installed on my laptop (I never developed an R package, so I never installed it). I checked and it seems I would need a pretty old version of XCode, because the latests are not compatible with OSX 10.12.
Fortunately, my laptop is recent enough that I can upgrade to the latest OSX version (Catalina). I'll try again when the update is complete and keep you posted!
