### 2020-06-03 samhforbes/PupillometryR#12 @samhforbes
Sorry for the fuss. If I had to guess from the bits and pieces of similar issues that people have, it looks like the compiler used in R 4.0.0 might have changed to needing XCode shipped with 10.13, but I can't see verification of this anywhere. An older version of R might also work, but I can't be certain.
