### 2020-06-05 samhforbes/PupillometryR#9 @Athanasiamo
This looks much neater in my opinion. Great you also added some badges indicating CRAN and CI status.

as a complete sidenote and something _not_ related to the review, but something you might consider, is using pkgdown to generate a site for your package. https://pkgdown.r-lib.org/articles/pkgdown.html
Your package is set up correctly, meaning there is little you have to do to get it working. For people looking for your vignette for instance, it might be more convenient to see it hosted online (though github pages).

But, this is purely a _suggestion_. As a developer though, I can say that I have often linked to the online vignettes rather than the vignettes shipped with the package. Particularly for developer updates not yet available on CRAN.
