### 2020-06-08 #2285 @oliviaguest
@samhforbes Just to keep this under control as there are a lot of issues on the code repo, can you close (button at the bottom) these, presuming they are addressed, please? And of course any others that are dealt with. 😊

* https://github.com/samhforbes/PupillometryR/issues/14
* https://github.com/samhforbes/PupillometryR/issues/11
* https://github.com/samhforbes/PupillometryR/issues/13
* https://github.com/samhforbes/PupillometryR/issues/10
* https://github.com/samhforbes/PupillometryR/issues/8
* https://github.com/samhforbes/PupillometryR/issues/7
* https://github.com/samhforbes/PupillometryR/issues/6
