### 2020-06-08 samhforbes/PupillometryR#9 @samhforbes
This is a great suggestion, and again, thanks for the advice. I'm going to leave this open temporarily as a reminder to devote some time to getting a (pretty) page going
