### 2020-06-08 samhforbes/PupillometryR#12 @paulinepalma
Hi Sam! No worries. I have now tested the developer version of your package on both Mac OS Catalina and on Windows 10 under the latest version of R (4.0.1), and everything went very smoothly! It is very possible the problem was with XCode, as I was directly prompted to download it once I opened R after updating the OS, and I had no problem afterwards.
