### 2020-06-09 samhforbes/PupillometryR#12 @samhforbes
Brilliant, thanks so much. I'm glad you got it going again - I suspect you aren't the last with this issue, so it's great to have it documented.
