+++ README.Rmd
@@ -91,9 +91,9 @@ This package has had suggestions, encouragement, and help from a number of peopl
 
 [1] Jackson, I., & Sirois, S. (2009). Infant cognition: Going full factorial with pupil dilation. *Developmental Science*, 12(4), 670-679. http://doi.org/10.1111/j.1467-7687.2008.00805.x
 
-[2] Allen, M., Poggiali, D., Whitaker, K., Marshall, T. R., & Kievit, R. (2018). Raincloud plots: a multi-platform tool for robust data visualization. *Wellcome Open Research*, 4, 1-41.
-doi:10.12688/wellcomeopenres.15191.1
+[2] Allen, M., Poggiali, D., Whitaker, K., Marshall, T. R., & Kievit, R. (2019). Raincloud plots: a multi-platform tool for robust data visualization. *Wellcome Open Research*, 4, 1-41.
+http://doi.org/10.12688/wellcomeopenres.15191.1
 
 [3] Ramsay, J.O., & Silverman, B.W. (1997). *Functional data analysis*. New York: Springer-Verlag.
 
-[4] van Rij, J., Hendriks, P., van Rijn, H., Baayen, R. H., & Wood, S. N. (2019). Analyzing the time course of pupillometric data. Trends in Hearing, 23, 233121651983248. https://doi.org/10.1177/2331216519832483
+[4] van Rij, J., Hendriks, P., van Rijn, H., Baayen, R. H., & Wood, S. N. (2019). Analyzing the time course of pupillometric data. *Trends in Hearing*, 23, 233121651983248. https://doi.org/10.1177/2331216519832483
