+++ README.md
@@ -127,13 +127,14 @@ factorial with pupil dilation. *Developmental Science*, 12(4), 670-679.
 <http://doi.org/10.1111/j.1467-7687.2008.00805.x>
 
 \[2\] Allen, M., Poggiali, D., Whitaker, K., Marshall, T. R., & Kievit,
-R. (2018). Raincloud plots: a multi-platform tool for robust data
+R. (2019). Raincloud plots: a multi-platform tool for robust data
 visualization. *Wellcome Open Research*, 4, 1-41.
-<doi:10.12688/wellcomeopenres.15191.1>
+<http://doi.org/10.12688/wellcomeopenres.15191.1>
 
 \[3\] Ramsay, J.O., & Silverman, B.W. (1997). *Functional data
 analysis*. New York: Springer-Verlag.
 
 \[4\] van Rij, J., Hendriks, P., van Rijn, H., Baayen, R. H., & Wood, S.
-N. (2019). Analyzing the time course of pupillometric data. Trends in
-Hearing, 23, 233121651983248. <https://doi.org/10.1177/2331216519832483>
+N. (2019). Analyzing the time course of pupillometric data. *Trends in
+Hearing*, 23, 233121651983248.
+<https://doi.org/10.1177/2331216519832483>
