+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: PupillometryR
 Type: Package
 Title: A Unified Pipeline for Pupillometry Data
-Version: 0.0.2.9000
+Version: 0.0.3
 Authors@R: 
         c(person("Samuel", "Forbes", email = "samuel.h.forbes@gmail.com", role = c("aut", "cre")),
         person("David", "Robinson", email = "admiral.david@gmail.com", role = "ctb"))
