### 2020-06-12 #2285 @samhforbes
I'm going to need to update the version because I've had to resubmit to CRAN (archived due to an older dependency that was fixed prior to peer review here).

Please let me know if there's anything else I can do
