### 2020-06-15 #2285 @oliviaguest
Hey all! 👋 @paulinepalma, @Athanasiamo, @samhforbes! Thank you to everybody for doing such a great job so far.

👉 Specifically for the reviewers, when they get a chance, check if their issues have been addressed and those that have are checked/ticked off above? 🌸

👉 For Sam, can you let me know which bits you are working on — again when you have the chance, no pressure. 🌼
