+++ NEWS.md
@@ -1,4 +1,11 @@
-### Update May 2020 PupillometryR 0.0.2
+# News
 
--   Minor patch to avoid breaks with new version of dplyr
--   Warning errors removed from ggplot2 when plotting
+## Update June 2020 PupillometryR 0.0.3
+
+* Documentation moved to vignettes
+* Dependency on package *spectral* removed due to problems with CRAN
+
+## Update May 2020 PupillometryR 0.0.2
+
+* Minor patch to avoid breaks with new version of dplyr
+* Warning errors removed from ggplot2 when plotting
