### 2020-06-15 samhforbes/PupillometryR#9 @samhforbes
It's pretty!
This is now up - currently at samforbes.me/PupillometryR
