+++ paper/paper.bib
@@ -65,13 +65,19 @@ url = {https://CRAN.R-project.org/package=fda }
     url = {https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=https://oraprdnt.uqtr.uquebec.ca/pls/public/gscw031?owa_no_site=314&owa_no_fiche=3&owa_bottin=}
 }
 
-@article{Geller2019,
-    title = {{GazeR: A package for processing gaze position and pupil size data}},
-    year = {2019},
-    journal = {PsyArXiv Preprints},
-    author = {Geller, Jason and Winn, Matthew B. and Mahr, Tristan and Mirman, Daniel},
-    doi = {10.31234/osf.io/gvcxb}
+@article{Geller2020,
+abstract = {Eye-tracking is widely used throughout the scientific community, from vision science and psycholinguistics to marketing and human-computer interaction. Surprisingly, there is little consistency and transparency in preprocessing steps, making replicability and reproducibility difficult. To increase replicability, reproducibility, and transparency, a package in R (a free and widely used statistical programming environment) called gazeR was created to read and preprocess two types of data: gaze position and pupil size. For gaze position data, gazeR has functions for reading in raw eye-tracking data, formatting it for analysis, converting from gaze coordinates to areas of interest, and binning and aggregating data. For data from pupillometry studies, the gazeR package has functions for reading in and merging multiple raw pupil data files, removing observations with too much missing data, eliminating artifacts, blink identification and interpolation, subtractive baseline correction, and binning and aggregating data. The package is open-source and freely available for download and installation: https://github.com/dmirman/gazer. We provide step-by-step analyses of data from two tasks exemplifying the package's capabilities.},
+author = {Geller, Jason and Winn, Matthew B. and Mahr, Tristian and Mirman, Daniel},
+doi = {10.3758/s13428-020-01374-8},
+file = {:Users/sam/Downloads/Geller2020{\_}Article{\_}GazeRAPackageForProcessingGaze.pdf:pdf},
+issn = {15543528},
+journal = {Behavior Research Methods},
+keywords = {R,eye-tracking,open science,pupillometry,visual world paradigm},
+publisher = {Behavior Research Methods},
+title = {{GazeR: A Package for Processing Gaze Position and Pupil Size Data}},
+year = {2020}
 }
+
 
 @book{Ramsay1997,
     title = {{Functional Data Analysis}},
