+++ paper/paper.bib
@@ -38,10 +38,10 @@
 }
 
 @misc{Ramsay2020,
-author = {Ramsay, J. O. and {Graves S.} and Hooker, G.},
-title = {{fda: Functional Data Analysis}},
-year = {2020}
-url = {https://CRAN.R-project.org/package=fda }
+  author = {Ramsay, J. O. and {Graves S.} and Hooker, G.},
+  title = {{fda: Functional Data Analysis}},
+  year = {2020}
+  url = {https://CRAN.R-project.org/package=fda }
 }
 
 
