### 2020-06-15 #2285 @samhforbes
I've just closed the last issue on the repository, which was to launch a pkgdown website as recommended by @Athanasiamo.

I've also updated the Geller reference as I've seen the paper has now come out in full, and Whedon preview suggests it compiles OK.
