### 2020-06-16 #2285 @paulinepalma
Yes! All of my comments have been addressed.

Le lun. 15 juin 2020, à 12 h 06, Olivia Guest <notifications@github.com> a
écrit :

> Thank you — amazing!
>
> @paulinepalma <https://github.com/paulinepalma> I assume this is all your
> requests/ideas addressed, right? 😊
>
