### 2020-06-16 d549da8
Update paper.bib
try to get names right
