+++ paper/paper.bib
@@ -27,7 +27,7 @@
     title = {{Safe and sensible preprocessing and baseline correction of pupil-size data}},
     year = {2018},
     journal = {Behavior Research Methods},
-    author = {Math{\^{o}}t, Sebastiaan and Fabius, Jasper and Van Heusden, Elle and Van der Stigchel, Stefan},
+    author = {Math{\^{o}}t, Sebastiaan and Fabius, Jasper and {Van Heusden}, Elle and {Van der Stigchel}, Stefan},
     number = {1},
     pages = {94--106},
     volume = {50},
@@ -38,7 +38,7 @@
 }
 
 @misc{Ramsay2020,
-  author = {Ramsay, J. O. and {Graves S.} and Hooker, G.},
+  author = {Ramsay, J. O. and Graves, S. and Hooker, G.},
   title = {{fda: Functional Data Analysis}},
   year = {2020},
   url = {https://CRAN.R-project.org/package=fda }
@@ -123,7 +123,7 @@ year = {2020}
     title = {{Analyzing the time course of pupillometric data}},
     year = {2019},
     journal = {Trends in Hearing},
-    author = {van Rij, Jacolien and Hendriks, Petra and van Rijn, Hedderik and Baayen, R. Harald and Wood, Simon N.},
+    author = {{van Rij}, Jacolien and Hendriks, Petra and {van Rijn}, Hedderik and Baayen, R. Harald and Wood, Simon N.},
     pages = {1--22},
     volume = {23},
     doi = {10.1177/2331216519832483},
@@ -134,7 +134,7 @@ year = {2020}
     title = {{Pupil dilation as an index of effort in cognitive control tasks: A review}},
     year = {2018},
     journal = {Psychonomic Bulletin and Review},
-    author = {van der Wel, Pauline and van Steenbergen, Henk},
+    author = {{van der Wel}, Pauline and {van Steenbergen}, Henk},
     number = {6},
     pages = {2005--2015},
     volume = {25},
