### 2020-06-16 7e86309
Update paper.md
consistency
