+++ paper/paper.md
@@ -28,7 +28,7 @@ In this way, the very nature of any pupillometric study demands careful planning
 The second such cause of the variable methods in pupillometric studies stems from the analytical decisions made in the processing of the data.
 Due to many researchers following either closed, in-house analytical pipelines, or open-source pipelines that require costly software, or particular equipment at some point in the analysis, there has been a lack of consensus on the analytical decisions to be made in processing.
 These decisions include how to filter and clean the data [@Jackson2009], decisions to be made on baselining the data [@Mathot2018], and whether to take time-course or aggregate pupil size forward as a variable [@vanRij2019].
-Thus PupillometryR aims to assist experimenters by providing a clear pipeline which is both free and open-source, and available for usage with most common brands of eye-tracker.
+Thus `PupillometryR` aims to assist experimenters by providing a clear pipeline which is both free and open-source, and available for usage with most common brands of eye-tracker.
 
 # Implementation
 
