### 2020-06-16 #2285 @oliviaguest
@openjournals/dev I set the DOI for the archive normally like this

> @whedon set 10.5281/zenodo.3897291 as archive

and then recompiled the PDF (a bunch of times by now) but the new generated PDF doesn't link to the "Archive" where it should (left column of first page). I assume this is a new/related bug? 😱
