### 2020-06-16 #2285 @xuanxu
@oliviaguest the `generate pdf` command prepares a preview of the pdf without adding the archive DOI. I think that is only done when compiling the final version after acceptance.
