### 2020-06-16 #2285 @oliviaguest
@openjournals/joss-editors please publish this! 😸
